\name{AMGET_package}
\alias{AMGET_package}
\title{Post-processing tool for ADAPT 5}
\description{
AMGET allows to simply and rapidly creates highly informative diagnostic plots for ADAPT 5 models. Features include data analysis prior any modeling form either NONMEM or ADAPT shaped dataset, goodness-of-fit plots (GOF), posthoc-fits plots (PHF), parameters distribution plots (PRM) and visual predictive check plots (VPC) based on ADAPT output.
}
\usage{AMGET()}
\author{Benjamin Guiastrennec <guiastrennec@gmail.com>}
\references{
AMGET CPT:PSP paper <http://www.nature.com/psp/journal/v2/n7/full/psp201336a.html>\cr
AMGET Users Guide (Attached to this package)\cr
ADAPT 5 User manual <http://bmsr.usc.edu/software/adapt/>}
\keyword{Pharmacokinetics}
\keyword{Pharmacodynamics}
\keyword{ADAPT 5}
\keyword{AMGET}
\keyword{Modeling}
