\name{AMGET}
\alias{AMGET}
\title{The AMGET function}
\description{This function is calling the user interface of AMGET, it does not require any argument}
\usage{AMGET()}
\author{Benjamin Guiastrennec <guiastrennec@gmail.com>}
\references{
AMGET CPT:PSP paper <http://www.nature.com/psp/journal/v2/n7/full/psp201336a.html>\cr
AMGET Users Guide (Attached to this package)\cr
ADAPT 5 User manual <http://bmsr.usc.edu/software/adapt/>}
\keyword{Pharmacokinetics}
\keyword{Pharmacodynamics}
\keyword{ADAPT 5}
\keyword{AMGET}
\keyword{Modeling}